+++ packages/inngest/package.json
@@ -202,7 +202,6 @@
     "@bufbuild/protobuf": "^2.2.3",
     "@inngest/ai": "0.1.2",
     "@jpwilliams/waitgroup": "^2.1.1",
-    "@standard-schema/spec": "^1.0.0",
     "@types/debug": "^4.1.12",
     "canonicalize": "^1.0.8",
     "chalk": "^4.1.2",
@@ -266,8 +265,7 @@
     "ts-jest": "^29.1.0",
     "ts-proto": "^2.6.1",
     "tsx": "^3.12.7",
-    "typescript": "5.8.2",
-    "valibot": "1.0.0-rc.3"
+    "typescript": "5.8.2"
   },
   "volta": {
     "node": "18.12.1",
