+++ packages/inngest/src/experimental.ts
@@ -1,6 +1,2 @@
 export { getAsyncCtx } from "./components/execution/als.js";
 export type { AsyncContext } from "./components/execution/als.js";
-
-export { channel, typeOnlyChannel } from "./components/realtime/channel.js";
-export { topic } from "./components/realtime/topic.js";
-export type { Realtime } from "./components/realtime/types.js";
