+++ packages/realtime/README.md
@@ -1,0 +1,1 @@
+# @inngest/realtime
