+++ packages/realtime/jsr.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "https://jsr.io/schema/config-file.v1.json",
+  "name": "@inngest/realtime",
+  "description": "",
+  "version": "0.0.0",
+  "include": ["./src/**/*.ts"],
+  "exports": {
+    ".": "./src/index.ts"
+  }
+}
