+++ packages/realtime/package.json
@@ -1,0 +1,61 @@
+{
+  "name": "@inngest/realtime",
+  "version": "0.0.0",
+  "description": "",
+  "main": "dist/index.js",
+  "publishConfig": {
+    "registry": "https://registry.npmjs.org"
+  },
+  "scripts": {
+    "test": "jest",
+    "build": "tsc",
+    "postversion": "pnpm run build",
+    "release": "node ../../scripts/release/publish.js && pnpm dlx jsr publish --allow-slow-types --allow-dirty",
+    "release:version": "node ../../scripts/release/jsrVersion.js"
+  },
+  "files": [
+    "dist"
+  ],
+  "exports": {
+    ".": {
+      "require": "./dist/index.js",
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
+  },
+  "keywords": [
+    "inngest",
+    "realtime"
+  ],
+  "homepage": "https://github.com/inngest/inngest-js/tree/main/packages/realtime#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/inngest/inngest-js.git",
+    "directory": "packages/realtime"
+  },
+  "author": "Inngest Inc. <hello@inngest.com>",
+  "license": "Apache-2.0",
+  "peerDependencies": {
+    "inngest": "^3.32.5"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.7.0",
+    "@types/debug": "^4.1.12",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.13.10",
+    "eslint": "^9.18.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "globals": "^15.14.0",
+    "inngest": "^3.32.5",
+    "jest": "^29.3.1",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^7.16.1",
+    "valibot": "1.0.0-rc.3"
+  },
+  "dependencies": {
+    "@standard-schema/spec": "^1.0.0",
+    "debug": "^4.3.4",
+    "zod": "^3.24.2"
+  }
+}
